requirements.txt
uvicorn
fastapi[standard]
fastapi-structured-logging
aiohttp
aiofiles
lxml
aiopath
load_dotenv
python-dotenv